+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # CHANGELOG
 
-## 0.X
+## Unreleased, TBD
 - Fixed iron data save issue [PR 122](https://github.com/NREL/H2Integrate/pull/122)
+- Added optional inputs to electrolyzer model, including curve coefficients and water usage rate.
+- Bug-fix in electrolyzer outputs (H2_Results) if some stacks are never turned on.
 
 ## 0.2 [7 April 2025]